refactor(cli): extract sessions command family into cli/commands/sessions.ts
